pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ministral-finetuner"
version = "0.1.0"
description = "A framework for fine-tuning Ministral-3B models locally"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "unsloth @ git+https://github.com/unslothai/unsloth.git",
    "datasets>=2.14.0",
    "trl>=0.7.0",
    "peft>=0.7.0",
    "accelerate>=0.25.0",
    "bitsandbytes>=0.41.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "rich>=13.7.0",
]

[project.scripts]
ministral-train = "cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-mock>=3.12",
]

[tool.setuptools.packages.find]
where = ["src"]
